% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{methods}
\name{Authentication,Ipfinder-method}
\alias{Authentication,Ipfinder-method}
\title{Get details for an Your IP address.}
\usage{
\S4method{Authentication}{Ipfinder}(theObject)
}
\arguments{
\item{theObject}{ipfinder class}
}
\value{
your ip data
}
\description{
Get details for an Your IP address.
}
\references{
\url{https://ipinfo.io/developers#full-ip-details}
}
