% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{methods}
\name{Call,Ipfinder-method}
\alias{Call,Ipfinder-method}
\title{Call to server}
\usage{
\S4method{Call}{Ipfinder}(theObject, path = NULL, format = NULL)
}
\arguments{
\item{theObject}{ipfinder class}

\item{path}{specific path of asn, IP address, ranges, Firewall,Token status}

\item{format}{available format \code{json} \code{jsonp} \code{php} \code{xml} and firewall formats}
}
\description{
Call to server
}
