% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{data}
\name{DEFAULT_API_TOKEN}
\alias{DEFAULT_API_TOKEN}
\title{The free token limited to 4,000 requests a day}
\format{An object of class \code{character} of length 1.}
\usage{
DEFAULT_API_TOKEN
}
\description{
The free token limited to 4,000 requests a day
}
\keyword{datasets}
