% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{data}
\name{DEFAULT_BASE_URL}
\alias{DEFAULT_BASE_URL}
\title{DEFAULT BASE URL}
\format{An object of class \code{character} of length 1.}
\usage{
DEFAULT_BASE_URL
}
\description{
DEFAULT BASE URL
}
\keyword{datasets}
