% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{data}
\name{DOMAIN_BY_PATH}
\alias{DOMAIN_BY_PATH}
\title{Domain By ASN, Country,Ranges path}
\format{An object of class \code{character} of length 1.}
\usage{
DOMAIN_BY_PATH
}
\description{
Domain By ASN, Country,Ranges path
}
\keyword{datasets}
