% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{data}
\name{DOMAIN_H_PATH}
\alias{DOMAIN_H_PATH}
\title{Get Domain IP history path}
\format{An object of class \code{character} of length 1.}
\usage{
DOMAIN_H_PATH
}
\description{
Get Domain IP history path
}
\keyword{datasets}
