% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{data}
\name{DOMAIN_PATH}
\alias{DOMAIN_PATH}
\title{Get Domain IP path}
\format{An object of class \code{character} of length 1.}
\usage{
DOMAIN_PATH
}
\description{
Get Domain IP path
}
\keyword{datasets}
