% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{data}
\name{FIREWALL_PATH}
\alias{FIREWALL_PATH}
\title{Firewall path}
\format{An object of class \code{character} of length 1.}
\usage{
FIREWALL_PATH
}
\description{
Firewall path
}
\keyword{datasets}
