% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{methods}
\name{getAddressInfo,Ipfinder-method}
\alias{getAddressInfo,Ipfinder-method}
\title{Get details for an IP address.}
\usage{
\S4method{getAddressInfo}{Ipfinder}(theObject, path)
}
\arguments{
\item{theObject}{ipfinder class}

\item{path}{IP address.}
}
\value{
ip data
}
\description{
Get details for an IP address.
}
\references{
\url{https://ipfinder.io/docs/#response-objects-details}
}
