% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{methods}
\name{getAsn,Ipfinder-method}
\alias{getAsn,Ipfinder-method}
\title{Get details for an AS number.}
\usage{
\S4method{getAsn}{Ipfinder}(theObject, path)
}
\arguments{
\item{theObject}{ipfinder class}

\item{path}{AS number.}
}
\value{
asn data
}
\description{
Get details for an AS number.
}
\references{
\url{https://ipfinder.io/docs/#response-object}
}
