% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{methods}
\name{getDomain,Ipfinder-method}
\alias{getDomain,Ipfinder-method}
\title{Get Domain IP}
\usage{
\S4method{getDomain}{Ipfinder}(theObject, path)
}
\arguments{
\item{theObject}{ipfinder class}

\item{path}{The API supports passing in a single website name domain name}
}
\value{
Domain IP data
}
\description{
Get Domain IP
}
\references{
\url{https://ipfinder.io/docs/#domain-ip-objects-details}
}
