% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{methods}
\name{getDomainBy,Ipfinder-method}
\alias{getDomainBy,Ipfinder-method}
\title{Get list Domain By ASN, Country,Ranges}
\usage{
\S4method{getDomainBy}{Ipfinder}(theObject, path)
}
\arguments{
\item{theObject}{ipfinder class}

\item{path}{The API supports passing in a single ASN,Country,Ranges}
}
\value{
list Domain data
}
\description{
Get list Domain By ASN, Country,Ranges
}
\references{
\url{https://ipfinder.io/docs/#domain-by-asn-objects-details}
}
