% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{methods}
\name{getDomainHistory,Ipfinder-method}
\alias{getDomainHistory,Ipfinder-method}
\title{Get Domain History IP}
\usage{
\S4method{getDomainHistory}{Ipfinder}(theObject, path)
}
\arguments{
\item{theObject}{ipfinder class}

\item{path}{The API supports passing in a single website name domain name}
}
\value{
Domain History data
}
\description{
Get Domain History IP
}
\references{
\url{https://ipfinder.io/docs/#domain-ip-history-objects-details}
}
