% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{methods}
\name{getFirewall,Ipfinder-method}
\alias{getFirewall,Ipfinder-method}
\title{Get Firewall data}
\usage{
\S4method{getFirewall}{Ipfinder}(theObject, path, formats)
}
\arguments{
\item{theObject}{ipfinder class}

\item{path}{AS number, alpha-2 country only.}

\item{formats}{list formats supported}
}
\value{
Firewall data
}
\description{
Get Firewall data
}
\references{
\url{https://ipfinder.io/docs/#response-details}
}
