% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{methods}
\name{getRanges,Ipfinder-method}
\alias{getRanges,Ipfinder-method}
\title{Get details for an Organization name.}
\usage{
\S4method{getRanges}{Ipfinder}(theObject, path)
}
\arguments{
\item{theObject}{ipfinder class}

\item{path}{Organization name.}
}
\value{
rang data
}
\description{
Get details for an Organization name.
}
\references{
\url{https://ipfinder.io/docs/#response-objects-details-3}
}
