% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{methods}
\name{getStatus,Ipfinder-method}
\alias{getStatus,Ipfinder-method}
\title{Get details for an API Token .}
\usage{
\S4method{getStatus}{Ipfinder}(theObject)
}
\arguments{
\item{theObject}{ipfinder class}
}
\value{
status data
}
\description{
Get details for an API Token .
}
\references{
\url{https://ipfinder.io/docs/#response-objects-details-2}
}
