% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder-package.R
\docType{package}
\name{ipfinder}
\alias{ipfinder}
\alias{ipfinder-package}
\title{The official Lua client library for the IPFinder.io}
\description{
IP address details (city, region, country, postal code, latitude and more ..).
ASN details (Organization name, registry,domain,comany_type, and more .. ).
Firewall by supported formats details (apache_allow, nginx_deny, CIDR , and more ..).
IP Address Ranges by the Organization name details (list_asn, list_prefixes , and more ..).
Get Domain IP (asn, organization,country_code ....).
Get Domain IP history (total_ip, list_ip,organization,asn ....).
Get list Domain By ASN, Country,Ranges (select_by , total_domain , list_domain ....).
}
\author{
Mohamed Ben rebia \href{mailto:mohamed@ipfinder.io}{mohamed@ipfinder.io}
}
