% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{data}
\name{RANGES_PATH}
\alias{RANGES_PATH}
\title{IP Address Ranges path}
\format{An object of class \code{character} of length 1.}
\usage{
RANGES_PATH
}
\description{
IP Address Ranges path
}
\keyword{datasets}
