% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfinder.R
\docType{data}
\name{STATUS_PATH}
\alias{STATUS_PATH}
\title{service status path}
\format{An object of class \code{character} of length 1.}
\usage{
STATUS_PATH
}
\description{
service status path
}
\keyword{datasets}
